requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.12.3
blinker==1.9.0
browser-cookie3==0.20.1
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
decorator==4.4.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.104.1
firebase_admin==7.1.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.0.1
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
instaloader==4.10.1
itsdangerous==2.2.0
Jinja2==3.1.2
lxml==5.1.0
lz4==4.4.5
MarkupSafe==3.0.3
moviepy==1.0.3
msgpack==1.1.2
multidict==6.7.0
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.4
Pillow==10.1.0
playwright==1.40.0
proglog==0.1.12
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.0.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
replicate==1.0.7
requests==2.31.0
rsa==4.9.1
selenium==4.16.0
shadowcopy==0.0.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.8.1
starlette==0.27.0
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.6.2
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
wsproto==1.3.2
yarl==1.22.0
python-telegram-bot==20.7